The loop is closed
The control is not correct yet, but at least the loop is closed. The
image processing node is returning a position of the ball on the board.
The controller takes that and finds the desired roll and pitch. The
serial_comm node writes the controls over a serial port. The arduino
reads that data and moves the platform to the desired position.