Some control
I made a node that would just publish arbitrary ball positions. I just
wanted to get the comms between the compy and arduino right. I made a
version of the arduino code that checks for start and end markers and
parses the roll and pitch from the message. It works at the moment, but
it is touchy. I have to send data to it at a rate of about 10hz right
now or else the data is unreliable.