Stable control
The platform has stable control of the ball right now. I took all print
statements out of the arduino code to try to speed it up, and I think
that worked. That let me speed up the frequency that the computer sends
commands to the arduino. Right now the control is stable, though it is a
slow response and it has oscillation. Increasing D gains at the moment
makes the platform very jittery.