Changes for new platform
The acryllic platform with green tape on the borders is on there now.
The ball_tracker node has been updates to work with that. The
platform isn't perfectly level, and since it's very smooth the ball
rolls very easily. I haven't gotten stable control with this setup yet.
I think I would need integral controllers as well as PD. The ball
bounces much better on this platform.