Best PID control so far
This is the best that I have save for now. I have had it better, but
then messed with it. The control has never been great. I don't know what
has to happen to improve it. One idea is to use AR tags to calibrate the
platform's angle. Also, maybe there's not just an offset on the angle,
but maybe there is a scaling between the commanded angle and the
platform's angle.